{"schema":"decodex/commit/1","summary":"Split operator status run projection","authority":"XY-1115"}
